requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=5.0,<6.0
beautifulsoup4>=4.12,<5.0
chromadb>=0.5,<1.0
ddgs>=9.0,<10.0
openai-whisper
psutil>=5.9,<7.0
pypdf>=5.0,<6.0
pyngrok>=7.0,<8.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
requests>=2.31,<3.0
sentence-transformers>=3.0,<4.0
torch
ytmusicapi>=1.8,<2.0
